add a piece interface for individual and all the pieces logic. Use that logic into the any type of piece. Optimice the piece controller and use piecefactory
