+++ src/pieces/IPiece.java
@@ -1,0 +1,9 @@
+package pieces;
+
+import board.Board;
+
+public interface IPiece {
+    void insertPiece(Board board, String color);
+
+    void insertAllPieces(Board board, String color);
+}
