+++ src/pieces/types/King.java
@@ -1,10 +1,14 @@
 package pieces.types;
 
+import board.Board;
+import pieces.IPiece;
 import pieces.Piece;
 import pieces.enums.Color;
 import pieces.enums.PieceType;
+import pieces.piece.PieceController;
 
-public class King extends Piece {
+
+public class King extends Piece implements IPiece {
     private final PieceType pieceType;
 
     public King(Color color) {
@@ -14,6 +18,25 @@ public class King extends Piece {
 
     public PieceType getPieceType() {
         return pieceType;
+    }
+
+    @Override
+    public void insertPiece(Board board, String color) {
+        System.out.println("---> MESSAGE: The KING is placed");
+
+        Color pieceColor = getColor(color);
+        int row = PieceController.getRow(color);
+        int emptyColumn = PieceController.findEmptyColumn(board, row);
+
+        board.getSquares(row, emptyColumn).setPiece(new King(pieceColor));
+    }
+
+    @Override
+    public void insertAllPieces(Board board, String color) {
+        Color pieceColor = getColor(color);
+        int row = PieceController.getRow(color);
+
+        board.getSquares(row, 4).setPiece(new Queen(pieceColor));
     }
 
     @Override
