+++ src/pieces/types/Knight.java
@@ -1,10 +1,13 @@
 package pieces.types;
 
+import board.Board;
+import pieces.IPiece;
 import pieces.Piece;
 import pieces.enums.Color;
 import pieces.enums.PieceType;
+import pieces.piece.PieceController;
 
-public class Knight extends Piece {
+public class Knight extends Piece implements IPiece {
     private final PieceType pieceType;
 
     public Knight(Color color) {
@@ -14,6 +17,25 @@ public class Knight extends Piece {
 
     public PieceType getPieceType() {
         return pieceType;
+    }
+
+    @Override
+    public void insertPiece(Board board, String color) {
+        System.out.println("---> MESSAGE: The HORSE is placed");
+
+        Color pieceColor = getColor(color);
+        int row = PieceController.getRow(color);
+        int emptyColumn = PieceController.findEmptyColumn(board, row);
+
+        board.getSquares(row, emptyColumn).setPiece(new Knight(pieceColor));
+    }
+
+    @Override
+    public void insertAllPieces(Board board, String color) {
+        System.out.println("---> MESSAGE: The HORSES are placed");
+
+        board.getSquares(PieceController.getRow(color), 1).setPiece(new Knight(getColor(color)));
+        board.getSquares(PieceController.getRow(color), 6).setPiece(new Knight(getColor(color)));
     }
 
     @Override
