+++ src/pieces/types/Pawn.java
@@ -1,10 +1,14 @@
 package pieces.types;
 
+import board.Board;
+import pieces.IPiece;
 import pieces.Piece;
 import pieces.enums.Color;
 import pieces.enums.PieceType;
+import pieces.piece.PieceController;
 
-public class Pawn extends Piece {
+public class Pawn extends Piece implements IPiece {
+    private static final int WIDTH = 8;
     private final PieceType pieceType;
 
     public Pawn(Color color) {
@@ -14,6 +18,29 @@ public class Pawn extends Piece {
 
     public PieceType getPieceType() {
         return pieceType;
+    }
+
+    @Override
+    public void insertPiece(Board board, String color) {
+        System.out.println("---> MESSAGE: A PAWN is placed");
+
+        Color pieceColor = getColor(color);
+        int row = (pieceColor == Color.WHITE) ? 1 : 6;
+        int emptyColumn = PieceController.findEmptyColumn(board, row);
+
+        board.getSquares(row, emptyColumn).setPiece(new Pawn(pieceColor));
+    }
+
+    @Override
+    public void insertAllPieces(Board board, String color) {
+        System.out.println("---> MESSAGE: The PAWNS are placed.");
+
+        Color pieceColor = getColor(color);
+        int pawnRow = (pieceColor == Color.WHITE) ? 1 : 6;
+
+        for (int i = 0; i < WIDTH; i++) {
+            board.getSquares(pawnRow, i).setPiece(new Pawn(pieceColor));
+        }
     }
 
     @Override
